# backend/documentation/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a PIXAGRAM-API documentation server.",
    "version": "1.0.0",
    "title": "PIXAGRAM-API Documentation",
    "termsOfService": "",
    "contact": {
      "email": "kanika.nadkarni9@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/post/user/{email}": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Find posts realted to user by user login-email",
        "description": "Returns all posts related to user",
        "operationId": "getPostOfUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "email of user to return all posts",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "404": {
            "description": "posts not found"
          }
        }
      },
      "post": {
        "tags": [
          "post"
        ],
        "summary": "Insert new post with image for a logged-in user with parameter as his login-email",
        "description": "",
        "operationId": "createPostofUser",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "logged-in email of an user",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "file to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/NewPost"
            }
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
...
    "/api/user/": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all users",
        "description": "",
        "operationId": "getAllUsers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by user e-mail",
        "description": "",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The user will be fetched by user e-mail.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/user/signup": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Registers new user in the system",
        "description": "",
        "operationId": "signupNewUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User registration/profile information fields to be saved",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserProfile"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/UserProfile"
            }
          },
          "400": {
            "description": "Registration failed"
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "loginUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User log-in credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginCredentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/LoginToken"
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/user/follow": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create User follower record in follows.",
        "description": "",
        "operationId": "followUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "email of person to be followed.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user1": {
                  "type": "string"
                },
                "user2": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Follower"
            }
          },
          "400": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/unfollow": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Unfollow perviously followed person.",
        "description": "",
        "operationId": "unfollowUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "email of person to be followed.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user1": {
                  "type": "string"
                },
                "user2": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/following/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all the users which the current user is following",
        "description": "",
        "operationId": "getFollowing",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Users email.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/followers/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all followers of a user",
        "description": "",
        "operationId": "getFollowers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Users email.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/followercount/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get number of followers of a user",
        "description": "",
        "operationId": "getFollowerCount",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Users email.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "followercount": {
                  "type": "integer"
                }
              }
            }
          },
          "404": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/activityFeed/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "getting posts of the people the current user is following in chronological order",
        "description": "",
        "operationId": "getActivityFeedPosts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Users email.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Posts": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "postTimestamp": {
                "type": "string",
                "format": "date-time"
              },
              "imagePath": {
                "type": "string"
              },
              "caption": {
                "type": "string"
              },
              "userEmail": {
                "type": "string"
              }
            }
          }
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "NewPost": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "imagePath": {
                "type": "string"
              },
              "caption": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          }
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "Users": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "email": {
                "type": "string"
              },
              "pswd": {
                "type": "string"
              },
              "fname": {
                "type": "string"
              },
              "lname": {
                "type": "string"
              },
              "bio": {
                "type": "string"
              },
              "dob": {
                "type": "string"
              },
              "gender": {
                "type": "string"
              },
              "country": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "profileimagePath": {
                "type": "string"
              },
              "coverimagePath": {
                "type": "string"
              },
              "visibility": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string"
        },
        "pswd": {
          "type": "string"
        },
        "fname": {
          "type": "string"
        },
        "lname": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "profileimagePath": {
          "type": "string"
        },
        "coverimagePath": {
          "type": "string"
        },
        "visibility": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "LoginCredentials": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "LoginToken": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "expiresIn": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "UserProfile": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "fname": {
          "type": "string"
        },
        "lname": {
          "type": "string"
        },
        "dob": {
          "type": "string",
          "format": "date-time"
        },
        "gender": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Follower": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "follow": {
          "type": "object",
          "properties": {
            "user1": {
              "type": "string"
            },
            "email": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}